requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\gen_rating_movies_app.py: 3
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating.py: 3
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_app.py: 3
Flask == 1.1.2

# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\gen_rating_movies_app.py: 13
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\genre_train.py: 7
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_app.py: 13
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_train.py: 7
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\revenue_train.py: 7
Pillow == 7.0.0

# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\gen_rating_movies_app.py: 5
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating.py: 5
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_app.py: 5
Werkzeug == 0.16.1

# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\gen_rating_movies_app.py: 11
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\genre_train.py: 5
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_app.py: 11
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_train.py: 5
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\revenue_train.py: 5
bs4 == 0.0.1

# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\gen_rating_movies_app.py: 16
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\genre_train.py: 10
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_app.py: 16
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_train.py: 10
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\revenue_train.py: 10
matplotlib == 3.3.1

# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\gen_rating_movies_app.py: 18,19,20,21,22
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\genre_train.py: 12,13,14,15,16
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating.py: 9
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_app.py: 18,19,20,21,22
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_train.py: 12,13,14,15,16
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\revenue_train.py: 12,13,14,15,16
nltk == 3.4.5

# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\gen_rating_movies_app.py: 14
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\genre_train.py: 8
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating.py: 8
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_app.py: 14
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_train.py: 8
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\revenue_train.py: 8
numpy == 1.18.5

# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\gen_rating_movies_app.py: 15
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\genre_train.py: 9
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating.py: 7
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_app.py: 15
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_train.py: 9
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\revenue_train.py: 9
pandas == 1.0.3

# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\gen_rating_movies_app.py: 10
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\genre_train.py: 4
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_app.py: 10
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_train.py: 4
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\revenue_train.py: 4
requests == 2.24.0

# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\gen_rating_movies_app.py: 27,28,29,31
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\genre_train.py: 21,22
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating.py: 12,13,14
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_app.py: 27,28,30
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_train.py: 21,22
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\revenue_train.py: 21,22,24,25,26
scikit_learn == 0.22.1

# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\gen_rating_movies_app.py: 17
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\genre_train.py: 11
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_app.py: 17
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_train.py: 11
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\revenue_train.py: 11
seaborn == 0.10.1

# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\gen_rating_movies_app.py: 12
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\genre_train.py: 6
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_app.py: 12
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\rating_train.py: 6
# C:\Users\sohana\Desktop\DSI\Module_3\NLP_rating\revenue_train.py: 6
wordcloud == 1.7.0

gunicorn == 20.0.4

yagmail == 0.14.245
